WIP: move ImgGen + sub-components to @vibes.diy/base for sandbox access
Moves ImgGen component, sub-components, hooks, and utilities from
use-vibes/base into vibes.diy/base (@vibes.diy/base package). This
package is already served via privateNpm and re-exported through
vibe-runtime, so sandbox apps can import { ImgGen } from "use-vibes".

Validated by experiment on jchris/test-use-vibes-privatenpm: re-exports
from @vibes.diy/base through vite-runtime work because the vite plugin
externalizes to the import map URL. use-vibes-base failed because its
deps (@clerk/react etc) aren't in the import map.

Build + lint pass. 6 test failures in ImgGen.test.tsx from hook rewrite.
