+++ use-vibes/base/package.json
@@ -12,6 +12,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@adviser/cement": "^0.5.34",
+    "@vibes.diy/base": "workspace:*",
     "@clerk/react": "^6.2.1",
     "@fireproof/core": "0.24.19",
     "@fireproof/core-keybag": "0.24.19",
