+++ use-vibes/tests/ImgGen.tracking.test.tsx
@@ -62,7 +62,7 @@ vi.mock("use-vibes", (actual) => {
 });
 
 // Import after mocks
-import { ImgGen } from "@vibes.diy/use-vibes-base";
+import { ImgGen } from "@vibes.diy/base";
 
 describe("ImgGen Document ID Tracking", () => {
   beforeEach(() => {
