+++ use-vibes/tests/package.json
@@ -59,6 +59,7 @@
   "dependencies": {
     "@react-router/dev": "^7.14.0",
     "@tailwindcss/vite": "^4.2.2",
+    "@vibes.diy/base": "workspace:*",
     "@vibes.diy/use-vibes-base": "workspace:0.0.0",
     "react-dom": "^19.2.5",
     "use-fireproof": "0.24.16",
