+++ vibes.diy/base/components/ControlsBar.tsx
