+++ vibes.diy/base/components/ImgGen.tsx
