+++ vibes.diy/base/components/ImgGenUtils/ImgGenDisplayPlaceholder.tsx
