+++ vibes.diy/base/components/ImgGenUtils/ImgGenDisplayUtils.ts
