+++ vibes.diy/base/components/ImgGenUtils/ImgGenError.tsx
