+++ vibes.diy/base/components/ImgGenUtils/ImgGenFileDrop.tsx
