+++ vibes.diy/base/components/ImgGenUtils/ImgGenPromptWaiting.tsx
