+++ vibes.diy/base/components/ImgGenUtils/overlays/DeleteConfirmationOverlay.tsx
