+++ vibes.diy/base/components/ImgGenUtils/overlays/ImageOverlay.tsx
