+++ vibes.diy/base/components/PromptBar.tsx
