+++ vibes.diy/base/hooks/image-gen/utils.ts
