+++ vibes.diy/base/index.ts
@@ -3,3 +3,21 @@ export * from "./components/icons/index.js";
 export * from "./components/index.js";
 export { cx, colors, semantic, stripesOverlay, gridBackground, getVibesGlobalCSS } from "./theme/index.js";
 export type { SemanticTheme } from "./theme/index.js";
+export { default as ImgGen } from "./components/ImgGen.js";
+export type { ImgGenProps } from "./components/ImgGen.js";
+export { useImageGen } from "./hooks/image-gen/index.js";
+export { ControlsBar } from "./components/ControlsBar.js";
+export { PromptBar } from "./components/PromptBar.js";
+export { ImgGenDisplay } from "./components/ImgGenUtils/ImgGenDisplay.js";
+export { ImgGenDisplayPlaceholder } from "./components/ImgGenUtils/ImgGenDisplayPlaceholder.js";
+export { ImgGenModal } from "./components/ImgGenUtils/ImgGenModal.js";
+export type { ImgGenModalProps } from "./components/ImgGenUtils/ImgGenModal.js";
+export { ImageOverlay } from "./components/ImgGenUtils/overlays/ImageOverlay.js";
+export { defaultClasses } from "./utils/style-utils.js";
+export type {
+  ImageDocument,
+  PartialImageDocument,
+  UseImageGenOptions,
+  UseImageGenResult,
+  ImgGenClasses,
+} from "@vibes.diy/use-vibes-types";
