+++ vibes.diy/base/package.json
@@ -11,10 +11,16 @@
   "dependencies": {
     "@adviser/cement": "^0.5.34",
     "@emotion/css": "^11.13.5",
-    "react": "^19.2.5"
+    "@fireproof/core-runtime": "0.24.19",
+    "@fireproof/use-fireproof": "0.24.19",
+    "@vibes.diy/use-vibes-types": "workspace:*",
+    "@vibes.diy/vibe-runtime": "workspace:*",
+    "react": "^19.2.5",
+    "react-dom": "^19.2.5"
   },
   "devDependencies": {
-    "@types/react": "^19.2.14"
+    "@types/react": "^19.2.14",
+    "@types/react-dom": "^19.2.3"
   },
   "repository": {
     "type": "git",
