+++ vibes.diy/vibe/runtime/img-gen.tsx
@@ -1,6 +1,9 @@
 import { VibeSandboxApi } from "./register-dependencies.js";
 import { isResErrorImageGen } from "@vibes.diy/vibe-types";
 
+// Re-export ImgGen component from @vibes.diy/base so sandbox apps can import { ImgGen } from "use-vibes"
+export { ImgGen, useImageGen } from "@vibes.diy/base";
+
 export let imageGen: (prompt: string) => Promise<string[]>;
 
 export function registerImageGen(vibeApi: VibeSandboxApi): void {
